Add files via upload
You are required to write a program that maintains a Playlist of favourite songs. The Playlist should
hold for each song, the song title, artist and year.
You will need Input.java, ArrayUtility.java and InsertionSort.java from the Library folder on Moodle.
You are required to implement a Song class, and a PlayList class with appropriate attributes and
methods. The main method should be in a class called PlayListTest.
You should develop the program one step at a time and keep a copy of working versions.
For steps 1 and 2, you can assume that there is space available in the array to add a song
object and that song titles will be unique.
Step 1 (functionality is worth 7 marks)
 create a Song class with suitable attributes to represent the song title, artist and year
 create a PlayList class that will maintain an array of Song objects
 create a main method in PlayListTest that instantiates a PlayList object and offers
options to add a song object and to display the details of the songs
 keep a copy of this version of the programme.
Step 2 (additional functionality is worth 10 marks)
 update the program to allow the song details to be sorted by year (in descending
order) and if there is more than one song for a year, then to sort by song title (in
ascending order). Song titles should not be case sensitive.
 keep a copy of this version of the programme
Step 3 (additional functionality is worth 8 marks)
 create two Junit classes. One to test the sorting of songs from Step 2 above and the
other to test the exception handling below.
 update the program to use exception handing to trap attempts to overfill the array
and to ensure that song titles are unique
 format the information to make the output more presentable